Title: 특정 구간의 원소 평균값, Time: 125ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240229/특정 구간의 원소 평균값/elemental-mean-value-for-a-particular-interval.java
@@ -23,7 +23,7 @@ public class Main {
                     cnt++;
                 }
 
-                double avg = (double) temp / cnt;
+                int avg = temp / cnt;
 
                 for (int start = k; start < k + i; start++) {
                     if (avg == list[start]) {
